Moved ModuleView-folder up a level.
